Fixed getWhen method in Action class.

+++ src/main/java/com/kalebdutson/app/Action.java
@@ -10,7 +10,7 @@ public abstract class Action {
     public Action(long when){
         this.when = when;
     }
-    public float getWhen() {
+    public long getWhen() {
         return this.when;
     }
 }
